Disabled movement when opening UIs for Fridge and Oven
Used a bool inside an if statement called interactable that applies to all floor tiles to prevent the character from moving when a UI is open. The boolean value exists in Move.cs, which is already on every FloorTile object. By using a foreach statement in FridgeUI.cs and OvenUI.cs, the interactable bool will be changed to false on MouseClick down. Upon exit it will be changed to true (edited in Fridge.cs and Stove.cs).